force context to fork for multiprocessing during cell extraction
fork is required here because otherwise we can not access the norm_function and the minmax_function for generating single cell images in the resulting workers. Also added additional check that if this code is run on windows threads is automatically set to 1 as "fork" context is not supported on windows.